src/requirements.txt
asgiref==3.7.2
bson==0.5.10
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.2
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3